refactor: move player name and outline color management to UsernameDisplay instead of Character
